+++ changelog.md
@@ -14,7 +14,10 @@
   - Wallet credits can be added to the shopping cart anywhere using a PageBuilder block.
   - An option has been added to display wallet credit amounts as buttons when purchasing credit.
 - #1445 AI: Create a new image using the Media Manager.
-- GMC: Added a setting to export products only for those for which GMC data is specified.
+- **GMC**
+  - Added a setting to export products only for those for which Google data is specified.
+  - When deleting rows in Google grid, delete the Google data assigned to the product.
+  - Consider the permission to update products.
 
 ### Improvements